+++ lw3/Streams/decorators/CDecryptionInputStream.h
@@ -5,13 +5,13 @@
 class CDecryptionInputStream : public CInputStreamDecorator
 {
 public:
-	CDecryptionInputStream(IInputDataStreamPtr&& stream, unsigned key);
+	CDecryptionInputStream(IInputDataStreamPtr&& stream, unsigned long key);
 
 	uint8_t ReadByte() final;
 	std::streamsize ReadBlock(void* dstBuffer, std::streamsize size) final;
 
 private:
-	void GenerateDecryptTable(unsigned key);
+	void GenerateDecryptTable(unsigned long key);
 	uint8_t DecryptByte(uint8_t byte) const;
 
 	std::vector<uint8_t> m_decryptTable;
